Support for binary JWT keys in HMAC algorithms.
The existing code in clj_jwt expects the JWT key for HMAC signing to be
provided as String. This only works when the key can be represented as a
string. The keys I am using cannot be represented as String as they are
fully binary.

With this patch the key can be provided as a byte array as well.